requirements.txt
closure
cssmin
Flask
flask-assets
Flask-Script
gmusicapi
pyxdg
# xspf
git+git://github.com/alastair/xspf.git#egg=xspf
